perf: fixing performance hotspots to get 1.4 million cycles per second.

+++ src/constants.rs
@@ -20,8 +20,8 @@ pub const TARGET_FRAME_DURATION: std::time::Duration =
 /// CPU frequency [Hz].
 pub const CPU_FREQ_HZ: usize = 4194304;
 /// CPU period [ns].
-pub const CPU_CLOCK_NS: u128 = (1000_000_000.0 / CPU_FREQ_HZ as f64) as u128;
+pub const CPU_CLOCK_NS: usize = (1000_000_000.0 / CPU_FREQ_HZ as f64) as usize;
 /// CPU cycles per frame.
-pub const CYCLES_PER_FRAME: u128 = (CPU_FREQ_HZ as f64 / 59.7) as u128;
+pub const CYCLES_PER_FRAME: usize = (CPU_FREQ_HZ as f64 / 59.7) as usize;
 /// Maximum number of sprites per line.
 pub const MAX_SPRITES_PER_LINE: usize = 10;
